More tests refactoring.
