+++ src/Tester/SerializationTests.cs
@@ -23,9 +23,13 @@ TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR TH
 
 
 using System;
+using System.Globalization;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using Orleans.CodeGeneration;
 using Orleans.Serialization;
+using UnitTests.GrainInterfaces;
 
 namespace UnitTests.General
 {
@@ -48,26 +52,26 @@ namespace UnitTests.General
             DateTime inputLocal = DateTime.Now;
 
             DateTime outputLocal = SerializationManager.RoundTripSerializationForTesting(inputLocal);
-            Assert.AreEqual(inputLocal.ToString(), outputLocal.ToString());
+            Assert.AreEqual(inputLocal.ToString(CultureInfo.InvariantCulture), outputLocal.ToString(CultureInfo.InvariantCulture));
             Assert.AreEqual(inputLocal.Kind, outputLocal.Kind);
 
             // UTC Kind
             DateTime inputUtc = DateTime.UtcNow;
 
             DateTime outputUtc = SerializationManager.RoundTripSerializationForTesting(inputUtc);
-            Assert.AreEqual(inputUtc.ToString(), outputUtc.ToString());
+            Assert.AreEqual(inputUtc.ToString(CultureInfo.InvariantCulture), outputUtc.ToString(CultureInfo.InvariantCulture));
             Assert.AreEqual(inputUtc.Kind, outputUtc.Kind);
 
             // Unspecified Kind
             DateTime inputUnspecified = new DateTime(0x08d27e2c0cc7dfb9);
 
             DateTime outputUnspecified = SerializationManager.RoundTripSerializationForTesting(inputUnspecified);
-            Assert.AreEqual(inputUnspecified.ToString(), outputUnspecified.ToString());
+            Assert.AreEqual(inputUnspecified.ToString(CultureInfo.InvariantCulture), outputUnspecified.ToString(CultureInfo.InvariantCulture));
             Assert.AreEqual(inputUnspecified.Kind, outputUnspecified.Kind);
         }
 
         [TestMethod, TestCategory("BVT"), TestCategory("Functional"), TestCategory("Serialization")]
-        public void SerializationTests_JObject()
+        public void SerializationTests_JObject_Example1()
         {
             const string json = @"{ 
                     CPU: 'Intel', 
@@ -82,10 +86,10 @@ namespace UnitTests.General
             Assert.AreEqual(input.ToString(), output.ToString());
         }
 
-        [Orleans.CodeGeneration.RegisterSerializerAttribute()]
-        internal class JObjectSerialization_Example1
+        [RegisterSerializerAttribute]
+        public class JObjectSerializationExample1
         {
-            static JObjectSerialization_Example1()
+            static JObjectSerializationExample1()
             {
                 Register();
             }
@@ -93,6 +97,7 @@ namespace UnitTests.General
             public static object DeepCopier(object original)
             {
                 // I assume JObject is immutable, so no need to deep copy.
+                // Alternatively, can copy via JObject.ToString and JObject.Parse().
                 return original;
             }
 
@@ -100,18 +105,116 @@ namespace UnitTests.General
             {
                 var input = (JObject)(untypedInput);
                 string str = input.ToString();
-                SerializationManager.SerializeInner(str, stream, typeof(string));
+                SerializationManager.Serialize(str, stream);
             }
 
             public static object Deserializer(Type expected, BinaryTokenStreamReader stream)
             {
-                var str = (string)(SerializationManager.DeserializeInner(typeof(string), stream));
+                var str = (string)(SerializationManager.Deserialize(typeof(string), stream));
                 return JObject.Parse(str);
             }
 
             public static void Register()
             {
                 SerializationManager.Register(typeof(JObject), DeepCopier, Serializer, Deserializer);
+            }
+        }
+        
+        [TestMethod, TestCategory("BVT"), TestCategory("Functional"), TestCategory("Serialization")]
+        public void SerializationTests_Json_InnerTypes_TypeNameHandling()
+        {
+            var original = new RootType();
+            var str = JsonConvert.SerializeObject(original, JsonSerializationExample2.Settings);
+            var jsonDeser = JsonConvert.DeserializeObject<RootType>(str, JsonSerializationExample2.Settings);
+            // JsonConvert fully deserializes the object back into RootType and InnerType since we are using TypeNameHandling.All setting.
+            Assert.AreEqual(typeof(InnerType), original.MyDictionary["obj1"].GetType());
+            Assert.AreEqual(typeof(InnerType), jsonDeser.MyDictionary["obj1"].GetType());
+            Assert.AreEqual(original, jsonDeser);
+
+            // Orleans's SerializationManager also deserializes everything correctly, but it serializes it into its own binary format
+            var orleansDeser = SerializationManager.RoundTripSerializationForTesting(original);
+            Assert.AreEqual(typeof(InnerType), jsonDeser.MyDictionary["obj1"].GetType());
+            Assert.AreEqual(original, orleansDeser);
+    
+            Console.WriteLine("Done OK.");
+        }
+
+        [TestMethod, TestCategory("BVT"), TestCategory("Functional"), TestCategory("Serialization")]
+        public void SerializationTests_Json_InnerTypes_NoTypeNameHandling()
+        {
+            var original = new RootType();
+            var str = JsonConvert.SerializeObject(original);
+            var jsonDeser = JsonConvert.DeserializeObject<RootType>(str);
+            // Here we don't use TypeNameHandling.All setting, therefore the full type information is not preserved.
+            // As a result JsonConvert leaves the inner types as JObjects after Deserialization.
+            Assert.AreEqual(typeof(InnerType), original.MyDictionary["obj1"].GetType());
+            Assert.AreEqual(typeof(JObject), jsonDeser.MyDictionary["obj1"].GetType());
+            // The below Assert actualy fails since jsonDeser has JObjects instead of InnerTypes!
+            // Assert.AreEqual(original, jsonDeser);
+    
+            // If we now take this half baked object: RootType at the root and JObject in the leaves
+            // and pass it through .NET binary serializer it would fail on this object since JObject is not marked as [Serializable]
+            // and therefore .NET binary serializer cannot serialize it.
+
+            // If we now take this half baked object: RootType at the root and JObject in the leaves
+            // and pass it through Orleans serializer it would work:
+            // RootType will be serialized with Orleans custom serializer that will be generated since RootType is defined
+            // in GrainInterfaces assembly and markled as [Serializable].
+            // JObject that is referenced from RootType will be serialized with JsonSerialization_Example2 below.
+
+            var orleansJsonDeser = SerializationManager.RoundTripSerializationForTesting(jsonDeser);
+            Assert.AreEqual(typeof(JObject), orleansJsonDeser.MyDictionary["obj1"].GetType());
+            // The below assert fails, but only since JObject does not correctly implement Equals.
+            //Assert.AreEqual(jsonDeser, orleansJsonDeser);
+
+            Console.WriteLine("Done OK.");
+        }
+
+        /// <summary>
+        /// A different way to configure Json serializer.
+        /// </summary>
+        [RegisterSerializer]
+        public class JsonSerializationExample2
+        {
+            internal static readonly JsonSerializerSettings Settings = new JsonSerializerSettings
+            {
+                NullValueHandling = NullValueHandling.Ignore,
+                TypeNameHandling = TypeNameHandling.All
+            };
+
+            static JsonSerializationExample2()
+            {
+                Register();
+            }
+
+            public static object DeepCopier(object original)
+            {
+                // I assume JObject is immutable, so no need to deep copy.
+                // Alternatively, can copy via JObject.ToString and JObject.Parse().
+                return original;
+            }
+
+            public static void Serialize(object obj, BinaryTokenStreamWriter stream, Type expected)
+            {
+                var str = JsonConvert.SerializeObject(obj, Settings);
+                SerializationManager.Serialize(str, stream);
+            }
+
+            public static object Deserialize(Type expected, BinaryTokenStreamReader stream)
+            {
+                var str = (string)SerializationManager.Deserialize(typeof(string), stream);
+                return JsonConvert.DeserializeObject(str, expected);
+            }
+
+            private static void Register()
+            {
+                foreach (var type in new[]
+                    {
+                        typeof(JObject), typeof(JArray), typeof(JToken), typeof(JValue), typeof(JProperty), typeof(JConstructor), 
+                    })
+                {
+                    SerializationManager.Register(type, DeepCopier, Serialize, Deserialize);
+                }
             }
         }
     }
